requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.3
redis==4.5.4
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.2.3
marshmallow==3.19.0
flask-cors==3.0.10 